Update README
# FiveM External by MND1337 | Version 1.0 BETA
![FiveM Mustache](https://i.imgur.com/bezTGSl.png)

> A Source Code That Chages The FiveM Comunity

## About
FlexBase Is A Fivem External Cheat Source With Native UI Remaked In A GUI Lib!

⚠️ Updating from 1.1 to 2.0 only possible with Whitelisted IPs and Bans


## Installation
To install FlexBase correctly, read the docs on [Watch The Video Of (.io)](mnd1337.dev/)

## Download
Download the latest version on [GitHub Releases](https://github.com/KuyaDotLua/Flex-Base/releases)

## License
```license
MIT License

Copyright (c) 2020 Thymon Arens (MND1337)

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
```

## Disclamer
FlexBase is a resource written to assist in finding possible cheaters. It does not mean that MND1337 makes cheating on your server impossible. Nor does it mean that FlexBase will be updated as soon as a new or different way of cheating is available. Project will be updated and modified as soon as there is time and reason to update the resource. It is allowed to rename, copy and distribute FlexBase as long as the original LICENSE is added to that project.